backend/requirements.txt
# Backend Python Dependencies

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0
psycopg2-binary==2.9.9  # PostgreSQL support (optional)

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
argon2-cffi==23.1.0

# Task Queue
celery==5.3.6
redis==5.0.1

# CLI
typer==0.9.0
rich==13.7.0
questionary==2.0.1

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
watchfiles==0.21.0

# Protocol-specific
pyftpdlib==1.5.9  # FTP server library
paramiko==3.4.0   # SSH/SFTP library

# SSL/TLS
certbot==2.8.0
cryptography==42.0.0

# System
psutil==5.9.8
python-daemon==3.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # For testing API

# Development
black==24.1.1
ruff==0.1.13
mypy==1.8.0
pre-commit==3.6.0
